HTBHFB-4 Only downloading cf cli if it does not already exist.

+++ ci_scripts/install_cf_cli.sh
@@ -4,8 +4,10 @@
 set -e
 
 echo "Installing cf cli"
-mkdir -p bin/
-cd bin
-wget "https://cli.run.pivotal.io/stable?release=linux64-binary&source=github" -q -O cf.tgz && tar -zxvf cf.tgz && rm cf.tgz
-cf --version
-cd ..
+if [[ ! -e bin/cf ]]; then
+    mkdir -p bin/
+    cd bin
+    wget "https://cli.run.pivotal.io/stable?release=linux64-binary&source=github" -q -O cf.tgz && tar -zxvf cf.tgz && rm cf.tgz
+    ./cf --version
+    cd ..
+fi
